Changes for adding state and state-msmt transition matrices
